+++ src/Mod/Fem/App/FemConstraint.h
@@ -59,11 +59,11 @@ protected:
 
 protected:
     /// Calculate the points where symbols should be drawn
-    const bool getPoints(std::vector<Base::Vector3f>& points, std::vector<Base::Vector3f>& normals) const;
-    const bool getCylinder(float& radius, float& height, Base::Vector3f& base, Base::Vector3f& axis) const;
-    Base::Vector3f getBasePoint(const Base::Vector3f& base, const Base::Vector3f& axis,
-                                const App::PropertyLinkSub &location, const float& dist);
-    const Base::Vector3f getDirection(const App::PropertyLinkSub &direction);
+    const bool getPoints(std::vector<Base::Vector3d>& points, std::vector<Base::Vector3d>& normals) const;
+    const bool getCylinder(double& radius, double& height, Base::Vector3d& base, Base::Vector3d& axis) const;
+    Base::Vector3d getBasePoint(const Base::Vector3d& base, const Base::Vector3d& axis,
+                                const App::PropertyLinkSub &location, const double& dist);
+    const Base::Vector3d getDirection(const App::PropertyLinkSub &direction);
 
 };
